Adjust line width based on stylus pressure
Use the pressure value in the event to adjust line width (brush size)
when drawing.  Previously, stroke will redraw all the points using
CoalescedEvents from the beginning until the pen is lifted, but this
will not work because CoalescedEvents do not contain pressure values,
so it makes sense to draw only from the last stroke ending position.