Draw prediction points using overlay canvas
Instead of drawing to a offscreen canvas and then re-drawing the entire
canvas over the original canvas, this draws just the prediction points
on a overlay canvas.  So there's no copying of pixels, it draws just the
predicted path on the overlay in real time.  This speeds up drawing and
also fix an issue where actual canvas and original canvas might have
different pixel rendering at different positions, so causing the line to
look slightly different when re-drawing the canvas.