Draw line between 2 points line using Path2D
This uses Path2D object to draw a smoother line from one width to
another width.  It first calculates the direction vector of the two
points, and calculate the angle, and then draw two semi-circle paths of
varying radius that adjusts based on the starting and ending points.
This allows a smooth transition of widh between strokes.