Fixed stroke showing gaps when drawing rapidly
This is due to the fact that when getCoalescedEvents() is used to draw
the stroke, the last point received from the coalesced event array is
not the same as the pointerMove event, so if we use the position
from the move event as the starting position for the next stroke, it
might have a gap.  The fix is to just re-use the last point in the
points array.