Refactor canvas implementation
De-coupled rendering logic to a separate renderer class so it can be
extended and re-used for other rendering implementations.  Created
BaseCanvas to abstract the canvas construction and touch event handling.
This will allow renders to be created in web worker more easily as web
worker do not have access to DOM elements and input events.