Fix rendering in points only mode
When drawing in points only mode, the rendering of coalesced points and
non-coalesed points are different, but the difference currently is not
stored in the history, so when you re-draw or redo history, it's not
rendering properly.  The points will need to be flagged when adding to
the paths before rendering when they are coalesced.