Added WebHID support
This patch allows the app to read/write color using WebHID when
navigator.usi is not present.  It will request the user to grant
permission to connect to the USI when it detect WebHID support in the
browser, which is supported in Chromium 0.89 or later.

NOTE: the kernel still need to be patched with HID access since Google
blocks all access to hidraw0 on chromebooks that exposes itself as a
mouse/keyboard.