Added read write status indicator
Currently there's a rare bug that occasionally reading or writing to the
stylus will fail, once this happen, it's not obvious that this will happen
unless there's a indicator to show the user that reading or writng has failed,
so adding a green check mark or a red error icon when error is caught.

Note: currently navigator.usi setPreferredColor do not throw exception,
only the webhid wrapper library HIDUSI will throw an exception.