Custom prediction using KalmanFilter
This patch implements a custom prediction algorithm using KalmanFilter
JS library, that samples the previous set of points and uses the
velocity and timestamps to predict the next set of predicted position.
The benefit using this is we have better control of what data is used to
feed into the KalmanFilter, for example, velocity only or with
acceleration.  Also added a select radio button to switch between chrome's
Chrome's prediction.